refactor: consolidate dbConnect
